+++ conf/config.yml
@@ -1,0 +1,2 @@
+conf:
+- userdn: "ou=users,dc=example,dc=com"
