+++ go.mod
@@ -1,0 +1,5 @@
+module ldap-self-service
+
+go 1.15
+
+require gopkg.in/yaml.v2 v2.4.0
