+++ static/form.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html:lang>
 
 <head>
     <meta charset="UTF-8" />
